fix: regenerate Prisma client to sync with schema
The schema is correct - relations exist between LiteLLM_VerificationToken
and LiteLLM_BudgetTable. The issue is that the generated Prisma Python
client is out of sync with the schema.

Schema has:
  LiteLLM_VerificationToken.LiteLLM_BudgetTable (relation)
  LiteLLM_BudgetTable.LiteLLM_VerificationToken (back-relation)

Error says:
  Field 'litellm_budget_table' does not exist

The Python field name (litellm_budget_table) should be generated from
the Prisma model name (LiteLLM_BudgetTable), but it's missing in the
generated client.

Solution:
  Clean regeneration of Prisma client:
  1. Stop LiteLLM
  2. Remove old generated client
  3. Generate new client from schema
  4. Verify generation
  5. Restart LiteLLM

Deploy:
  cd provision/ansible
  ansible-playbook -i inventory/test/hosts.yml site.yml --tags litellm_configure

  Then run:
  ssh root@10.96.201.208 /usr/local/bin/regenerate-prisma-client